Filter table rows by visible map markers
- Check if the Marker's produced g element is within bounds of map container viewport
- Filter table rows whose combinedRows are marked as visible
- Add configuration for element id's given a React component name
- Add extra types, like index and visibility
- Allow filtered rows to alternate in color
- Refactor functions into appropriately named utility files and rewrite some predicates
- Implement Luke's code review suggestions
- Move all components except App into new ./src/components dir
