+++ src/components/MapChart.tsx
@@ -1,27 +1,26 @@
 import React from "react";
-import ZoomControls from "./ZoomControls";
+import ZoomControls from "../components/ZoomControls";
 import { ComposableMap, ZoomableGroup } from "react-simple-maps";
-import * as Config from "./Config";
-import * as Functions from "./Functions";
-import * as Types from "./Types";
+import { handleMoveEnd } from "../utils/EventHandlers";
+import { createGeographies, createRowMarkers } from "../utils/Renderers";
+import * as Config from "../utils/Config";
+import * as Types from "../utils/Types";
 
 const MapChart = (props: Types.MapChartProps): JSX.Element => {
   const [state] = props.stateManager;
   return (
     <div className="mapchart">
       <ZoomControls stateManager={props.stateManager} />
-      <ComposableMap projectionConfig={{ scale: 200 }}>
+      <ComposableMap id="container" projectionConfig={{ scale: 200 }}>
         <ZoomableGroup
           zoom={state.mousePosition.zoom}
           center={state.mousePosition.coordinates}
-          onMoveEnd={(newPosition) =>
-            Functions.handleMoveEnd(newPosition, props.stateManager)
-          }
+          onMoveEnd={handleMoveEnd(props.stateManager)}
           minZoom={Config.minZoom}
           maxZoom={Config.maxZoom}
         >
-          {Functions.createGeographies(props.stateManager)}
-          {Functions.createRowMarkers(props.stateManager)}
+          {createGeographies(props.stateManager)}
+          {createRowMarkers(props.stateManager)}
         </ZoomableGroup>
       </ComposableMap>
     </div>
