+++ src/components/ZoomControls.tsx
@@ -1,24 +1,25 @@
 import React from "react";
+import { displayDebuggingFeatures } from "./../utils/Renderers";
+import { handleZoomIn, handleZoomOut } from "./../utils/EventHandlers";
 import { ReactComponent as ZoomIn } from "./zoom-in.svg";
 import { ReactComponent as ZoomOut } from "./zoom-out.svg";
-import * as Config from "./Config";
-import * as Functions from "./Functions";
-import * as Types from "./Types";
+import * as Config from "./../utils/Config";
+import * as Types from "./../utils/Types";
 
 const ZoomControls = (props: Types.ZoomControlProps) => {
   const [state] = props.stateManager;
   return (
     <div className="controls-wrapper">
-      {Functions.displayDebuggingFeatures(true, state.mousePosition)}
+      {displayDebuggingFeatures(true, state.mousePosition)}
       <div className="controls">
         <button
-          onClick={() => Functions.handleZoomIn(props.stateManager)}
+          onClick={handleZoomIn(props.stateManager)}
           style={{ background: Config.defaultMarkerColor }}
         >
           <ZoomIn />
         </button>
         <button
-          onClick={() => Functions.handleZoomOut(props.stateManager)}
+          onClick={handleZoomOut(props.stateManager)}
           style={{ background: Config.defaultMarkerColor }}
         >
           <ZoomOut />
